Fix SelectedVariantChooser in the Gradle Daemon.
When opening a project for the very first time in Studio that used AGP
3.2.x without Kotlin, sync would fail with Unsupported method:
AndroidProject.getDefaultVariant().

SelectedVariantChooser runs in two contexts, inside Studio with the
'safe' models, and in Gradle using models directly from the Android
Gradle Plugin.

When running in Gradle, it must handle recently added model methods to
potentially throw UnsupportedMethodException when running with older
Android Gradle Plugins.

Bug: 127702568
Test: Next change to expand the snapshot tests to include AGP 3.3.2.
Change-Id: I327021b1353d56ac8ecf1209269970e593e54890