Track observable types and classes
We are tracking three kinds of observables in this CL:
1. observable primitives in the likes of ObservableInt, ObservableField, etc.
2. observable collections. only 2 exist: ObservableArrayMap and ObservableArrayList.
3. user defined classes that directly or indirectly (through inheritance) implement Observable.

For #3, although there are utility functions that determine whether a PsiType is an inheritor of another type. AFAICT they only work for java and not for KtElement (Kotlin). Therefore, I opted to stick with the strategy that I used for #1 and #2, which is to maintain a hardcoded string "Observable" for comparisons.

Test: TrackObservableTest
Bug: 123029035
Change-Id: I62a9d68ff018376ec0b916320a798a6e16ca3779