Look for build files in base path for root module
When we look for a module's build file while the Gradle facets are not
available we assume that the build file and the module file will be in
the same folder, but it is possible that this is not the case, for
example, com.intellij.platform.PlatformProjectConfigurator#configureProject
creates the iml in <project_path>/.idea/<module_name>.iml but the
project still has the build file at <project_path>/build.gradle.

With this change, if the facets are still not available, we try to get the
project build file from the base dir, and if not posible then where the iml
file is located.

Bug: 126802162
Test: Covered by testGetGradleBuildFileFromModuleWithoutGradleFacet, added
a couple of new tests (one for project module and other for non project
module).

Change-Id: I62f5094c9ad9d390879a39bba2667e44e85ef54a