[Gradle Build Model] Fix an issue where it was possible to OOM
While resolving references we kept a set of GradlePropertyModels in
order to stop repeating the resolution. Since the GradlePropertyModels
are recreated on each iteration of the loop, and the comparison was done
via references, this could never actually detect anything.

This change stores the full name of the property instead which should
fix this case, preventing the constant creation of GradlePropertyModels
and correctly terminating the loop.

Fixes: 126952745
Test: do not know of any specific use case that trigger this
Change-Id: I344a4c825f69b5c93e8568dcae258232cabf3676