Fix overzealous optimization in DefaultObjectExpander
DefaultObjectExapnder overrode the default implementation of
getChildForLabel (which iterates over all of the outgoing edges and
checks their labels) to just reflectively access the field, and look up
the corresponding node in the map.

This behavior is OK while the graph is being expanded (since application
threads are paused), but is potentially incorrect during the final
analysis - the field may have changed value since the graph was
expanded. This seemed to be particularly problematic for
LockFreeCopyOnWriteArrayList#array.

This change removes the override so the default behavior is used. The
performance impact turns out to be minimal.

Change-Id: I6246351ea9b44d77345692c16f268c6ccf416129