Remove failing sanity check from create layout action
When we execute the "Create layout resource file" action,
we present the user with a dialog that allows them to specify the root
tag to use for the new layout. We suggest a default root tag, which is
either LinearLayout or ConstraintLayout, depending on whether the module
containing the new file depends on constraint layout. We also generate a
list of valid root tags that we use to give autocomplete suggestions
if the user wants to enter a non-default root tag.

Currently, we sanity check the default root tag selection by asserting
that it's in the list of all possible root tags. Unfortunately, some
non-Gradle projects don't set up the modules with all the right
dependencies. The project system APIs may give a different (more
reflective of the actual dependencies specified to the build system)
result than the module dependencies as known by IntelliJ. In such cases,
ConstraintLayout doesn't show up in the list of allowed root tags and the
assert will fail, preventing the "Create layout resource file" from doing
anything.

Since there's no harm in using ConstraintLayout in such cases (it won't
resolve correctly in the XML editor yet, but the app build/deploys
correctly), this CL resolves the issue by just removing the assert.

Bug: 112834556
Test: CreateMultiRootResourceFileActionTest
      + BlazeModuleSystemExternalDependencyIntegrationTest
      + manually added a new layout in such a module
        with ConstraintLayout as the root tag

Change-Id: I81cdee90bb14b95b05f7badeaadc1cb683e38508