Fix inconsistent reference test assertion
In IntelliJ, a PsiReferenceProvider takes a target PsiElement
and is expected to return any matching PsiReferences that the
source element resolves to.

A PsiReference, itself, wraps an underlying element. And it
turns out there is an (undocumented?) rule / expectation that
the element passed into the PsiReferenceProvider will also be
the element wrapped by the PsiReference instances it returns.

However, data binding has two PSI concepts: an `id` and a
`refExpr`, where a `refExpr` is essentially a qualified ID,
e.g. "a.b.c.d.value" is part prefix ("a.b.c.d") and part
id ("value")

Due to ambiguity, an expression like "value" is treated
initially like a "refExpr" but it ultimately delegates its
work to the "id" reference provider. This caused the input
element to be type "PsiDbRefExpr" but the wrapped PsiReference
element type to be type "PsiDbId".

The code for data binding's reference contributor has been
reworked to restore consistency and satisfy IntelliJ's
assertion.

Fixes: 126956711
Test: Covered by existing tests
Change-Id: Ied3061a828d96921de1b68c36ae1fcdf3131437d