[Gradle Sync] Persist SyncIssues in SyncIssueRegister until next sync
This change ensures that the SyncIssues from the last sync can always be
obtained from the SyncIssueRegister. We require this for tests since it
is sometimes possible for sync issues to fail the sync without the any
AndroidModuleModels populated (e.g SDK configuration failure). In this
case there is currently no way of accessing these SyncIssues.

We add error logs when sync issues are added after sync has
finished or when sync issues are read before sync has finished.

Test: existing
Bug: N/A
Change-Id: I9ea01c53bcbb99c26c297143c968c665e9267c2d