[Gradle Sync] Enable new sync for non-MPP Kotlin projects
The KotlinModelBuilder in the Kotlin Gradle plugin has a bug where it
will crash when there are no Kotlin compile tasks. This is the case
when the SDK configuration fails in the Android Gradle Plugin.

To work around this we first try to request the GradlePluginModel, if
this model is present and there is a no variants present then we
don't request the KotlinProject. We also remove the KotlinProject from
the models we request as otherwise it will crash when we try and request
it later, this is safe to do since we are guaranteed that sync will fail
with no models and so the absence of the KotlinProject will not break
anything.

Bug: 79581505
Test: updated
Change-Id: I5321e3c3849a58e10ea2abfd4dc01fae248736b4