Extract sync time tracking from ManifestInfo.ManifestFile
Currently, each ManifestInfo.ManifestFile instance registers a sync
listener that lasts the lifetime of its corresponding AndroidFacet
in order to track the timestamp of the last completed sync. This is
could be done with a stand-alone project-wide listener, which would not
only cut down the number of active listeners, but also remove some of
the clutter from ManifestInfo.ManifestFile.

This CL resolves the issue by replacing the sync listening logic in
ManifestInfo.ManifestFile with a call to the new SyncTimestamp, a
project component responsible for tracking sync times.

Bug: N/A
Test: ManifestInfoTest + SyncTimeListenerTest
Change-Id: I25b6476aac6f1c96c34405d7985a9e972c183cd0