Add cpu capture memory test to perfgate
This change allows us to measure the highwater mark and
memory used when loading a 10s perfetto and atrace. The
framework also allows us to do the same with a simple
perf capture.
The perfetto and atrace captures use the same arguments,
recorded on the same app while the app was in the same state.
This was done to help ensure the two captures were not significantly
different from eachother.

Test: Adding perfgate test.
Bug: None
Change-Id: Iabac42f5dddf8e97cfbc5fc32794110672e9bea5