Refresh library roots after library setup.
Library roots are generated by Gradle, and they're not updated
to Virtual File system unless file system is refreshed.
This change forced refresh for the missed library roots as part
of library setup.

Besides, update ArtifactsByConfigurationModuleSetupStep so that
all of setup steps use the common API in ModuleDependenciesSetup
for external libraries. The name of local jar dependency now have
prefix "Gradle: ", which is consistent with other types of modules.

Bug: 126740908
Bug: 126795724
Bug: 128043536
Test: included.
Change-Id: Ic0d283b91d4affeb1ff0a00967f225e30b30fef8