Dispose ResourceFolderRepositories with the project
Recently we stopped disposing them multiple times, but I missed the fact
that ResourceFolderRegistry didn't actually handle disposal at all.
After this change ResourceFolderRegistry becomes disposable (being a
project service, it's automatically tied to the project) and registers
itself as the parent of all created repositories. Also invalidates the
cache on disposal, just to be sure.

To make this work, I had to change ModuleResourceRepository slightly, to
introduce the contract that every repository returned from
`forMainResources` needs to be registered with the disposer. This
affects only legacy projects and tests, and should not change much since
MultiResourceRepository already has a fast path for a single child.

Bug: None yet, because ResourceFolderRepository does nothing on disposal,
     but will start soon.
Test: ResourceRepositoryManagerTest.java
Change-Id: I9d0dc5ffb61c3ad11be5361f69af93053d747a42