NDK side-by-side: Add error handler that will set ndk.dir
When NDK SxS is enabled, ndk.dir is not set automatically before the sync
begins. This adds an error handler and corresponding hyperlink to set the
value of ndk.dir in local.properties when this happens.

For parity, also do the same for CMake (set cmake.dir if it exists).

This is to support older plugins that can't find SxS NDK installs on their
own.

ndk-sxs-3.2.1

Bug: 80478236
Test: Tested with Gradle Plugin verison 3.2.1
Change-Id: I8a33e869e340a0790a030e9b47fafcaa7737e2cb