Split up ManifestInfo.ManifestFile#refresh
ManifestInfo.ManifestFile is responsible for maintaining a cached view
of an Android module's merged manifest. Currently, the logic for
determining whether the cached merged manifest DOM is valid is
lumped together with the long-runnning code for updating the manifest
in a single refresh() method. This makes it impossible for us to quickly
determine whether the merged manifest DOM is out of date in one thread
and then offload the expensive recalculation to a background thread.

This CL resolves the issue by splitting up the refresh method: the logic
for kicking off a merge and extracting relevant data from the result has
been moved into ManifestInfo.ManifestFile#create, and the logic for
determining whether an instance of ManifestInfo.ManifestFile is
still valid has been moved to a new isUpToDate() method. Since both
methods need to know which files contribute to the merged manifest, the
logic for computing the set of relevant files has been extracted to the
new MergedManifestContributors class to avoid code duplication.

An important difference after this change is that
ManifestInfo.ManifestFile is now an immutable data object. If callers
need to refresh the view of the merged manifest because it's out of
date, they must create a new instance (since it's no longer possible to
update the state of the existing one).

Bug: 128854237
Test: ManifestInfoTest + MergedManifestManagerTest
Change-Id: I0335a1ea072b453024e9cf0cf6ed214b5018ee80