Specify no undo before setting the initial value in editors.
The Nav editor (and the old property editor) was specifying no
interest in undo events AFTER the initial value was set.
Setting of the initial value created an undo event which could
be undone after the user has edited the field.
That is a problem because the undo record doesn't match the state
of the editor => errors.

Fixes: 123648277
Test: N/A
Change-Id: I73072d73cbaca762648418d3803fed53bf50f396