Extract ManifestInfo.ManifestFile to MergedManifestInfo
ManifestInfo.ManifestFile is a misleading name, since the object
includes auxiliary information about the result of computing the
merged manifest (logs, actions taken by the merger), and may not
even hold a reference to the manifest if the merge was unsuccessful.
At the same time, ManifestFile sits inside the ManifestInfo class,
whose only other memebers are static utilities that are used by
ManifestFile exclusively.

This CL tidies things up a bit by renaming ManifestFile to
MergedManifestInfo and making it the top-level class. The other static
members of the old ManifestInfo class now belong to MergedManifestInfo.

Test: existing (there are no functional changes)
Bug: 128854237
Change-Id: Id389c5cabf1ae557f3ab55495d4f26c4b66e1e6f