Fix conflictException in NavigationSchema
If the project contains two classes with the same fully
qualified name we can get a conflictException in buildCacheKeys.
objects with the same qualified name. The actual class type is
ClsClassImpl which compares by reference, while TypeRef
compares by qualified name.

Move the TypeRef for the navigator class into the cache object
and convert the dictionary to a list. This will prevent any
conflicts.

Fixes: 128030918

Test: NavigationSchemaTest
Change-Id: I6429a45cdbcc761173a1be3831945198556b3cb0