Remove obsolete quickfix logic for unresolved dependencies.
Nowadays we should not use nor offer artifacts from SDK m2 repo,
and suggest gMaven instead, as this is the only source of truth
and SDK m2 repos are outdated anyway. This CL changes the
unresolved dependency tracker logic to use the gMaven quickfix
for a wider set of dependencies, and cleans up the the code to
remove the usages of the obsolete quickfixes.

Test: Existing (w/ amendments)
Bug: 128825869
Change-Id: Ic609dd9600c161a26828be390c278fe922b03e2f