Add GradleDistributionInstallErrorHandler
When running Gradle tasks from a project that has a wrapper, the
wrapper first tries to check if the Gradle distribution is
installed and if not then downloads the zip file to the specified.
If the zip file already exists, the wrapper just tries to open it
instead of downloading it. If this file is corrupted, then Android
Studio would assume that the problem is in a file from the Gradle cache
and show a message to re download dependencies and try again, but this
does not solve the problem since Gradle cannot be run.

This CL adds GradleDistributionInstallErrorHandler to detect this issue.
When detected, an error messsage is shown, saying that the distribution
could not be installed and that the zip file may be corrupted. It also
offers a quickfix to delete the zip file (with a confirmation dialog).

Bug: 128914973
Test: Added tests to check the error is detecetd.
Change-Id: I7c41def91e749c26042cadceca6c18013aef7c61