Make template icons scale aware
In order to always have the best quality scaling of icons in the wizard,
we make sure to only manipulate them in a scale aware way.
Icons in the wizard need to be able to have their size set, as well as
being cropped. This creates a new class of icons, TemplateIcon, which
implement those two properties in a scale aware way.

Bug: 128689581
Test: checked visually on Windows with various scalings
Change-Id: Ia2472de46d10a6986d8b79e7caa6c769218dde28