Add Windows Defender system health check
This change adds collection of diagnostic information about the status
of Windows Defender, whenever a project is opened on Windows. It first
checks whether realtime scanning is enabled, and if so, it checks
whether certain important paths (project directory, gradle cache, etc)
are covered by the list of exclusions. If scanning is enabled and not
everything is excluded, a notification balloon will be shown with a link
for more information. Windows Defender settings are not automatically
adjusted by Studio.

Bug: 127513071
Change-Id: I1a941ed6110606603118b28c4cf1046bee43d17d