[resize] Remove unavailable layer from SimplerResize
If we're not swapping files, the unavailable layer no longer makes
sense. We allow any files to be of any size. The starting position is
off (because it no longer represent current file as white and gradient
cyan over). Only thing it might be useful is to know the starting
position?

It also improves the rendering slightly.

Bug: 129346896
Test: Manually tested while flag is on
Change-Id: I1bffa2f6073f60a7aa6f59f61ffc12d8e06d587b