Add an exception for support-annotations.
To workaround a bug in Jetpack treat the support-annotations as
a stable library instead of one of the special groups that need to
be the exact same version.

Modify the dependencies of appcompat-v7 and recyclerview-v7 to have
dependencies on support annotations with different version numbers.
The test: testAddSupportDependencyWithMatchInAppModule will now
fail unless we special case support-annotations.

Fixes: 129408604
Bug: 129067099
Test: Added unit test.
Change-Id: I4e0522cf336884b6e3d21fb765e64fdd0c6f08ac