Fixes multiple tabs/stop button numbers issue when swapping
Currently, when swap happens, we swap on a new ProcessHandler instead of the
original ProcessHandler. This causes a disconnect and reset of the output
Content in the ToolWindow. It also results in multiple tabs being created in
the (currently) Run ToolWindow, which is the proxy that determines the presence
of the "number" under the Stop button.

This CL fixes/hacks all this so that swapping happens in the same
ProcessHandler as the existing running app. When Running or Debugging, it
destroys the old RunContentDescriptor and creates a new one.

Fixes: 126949874
Fixes: 129411786
Fixes: 129411867
Test: Run/Swap and checked that there is only the intended tabs open
Change-Id: I19c055f0616d6cd7a7b4395aba0da8ed9dda6cf0