Look for "tools" package in the rest of SDK directory.
Currently SdkToolsVersionSetupStep will only check for the tools package
expecting it to be at $SDK/tools directory, but if it has been moved
somewhere else (for example $SDK/tools-non-default-dir) the installer
called by the pop-up will silently fail because the package is actually
already installed.

So, if the VersionCheck.isCompatibleVersion(...) check fails, we
fallback to use the RepoManager to look-up all the packages in the user
SDK directory.

Important note, this invocation of RepoManager.load(...) forces the
Downloader to null, to totally avoid any attemp to using the network. We
do this, because right here we only want to check if the package is
present/installed, not actually trigger any logic to install it.

Bug: 129349771
Test: SdkToolsVersionSetupStepTest
Change-Id: Id958b3a2cf9a2f1831492d6f830649173fa836ad