Ignore generated data binding "Binding" base classes
These "Binding" classes are already handled by
`LightBindingClass` instances, so the generated code
and the in-memory code clash. We were able to fix this
in the Java case by prioritizing `LightBindingClass` over
generated code, but the Kotlin plugin didn't seem to
expose this functionality, always given the generated
code preference.

By ignoring the generated code, we avoid issues with
inconsistent navigation destinations (now, we always
navigate to the xml file) and inconsistent code
completions (now, we always use the light binding class
as the source of truth for code completions).

Thanks to the design of data binding, debugging still
works, as stepping in will take users to the "Impl"
version of the binding, and we do not ignore the "Impl"
generated files.

Fixes: 123531078
Fixes: 123379503
Fixes: 119832794
Fixes: 123312680
Test: Added
Change-Id: I0cd215710ce6a00e94b035c0e0bfecd515a065e4