Migrate CreateAPKProjectTest.createProjectFromLocallyBuiltApk into Bazel
There are a few additional changes to the test to let the test be able
to run within Bazel. The IDE sets up an APK debug projects directory
under the user's home directory. When running within Bazel, the user is
set to the nobody user, which has no home directory by default. This
causes the test to fail.

Under a typical environment, the user will have a home directory, so
this test does not reflect a user's environment as accurately as it
should. In order to give the the test a more accurate test environment,
the test runs as fake root, and a temporary directory is mounted on the
home directory, which is usually /root on a typical Linux system. Since
this mount is set up under a mount namespace in the Bazel sandbox, this
test will be unable to affect the mount namespace or the /root directory
outside of the sandbox. With this temporary home directory, the test
will now be able to run.

Bug: 74079626
Test: bazel test //tools/adt/idea/android-uitests:LocalApkProjTest
Change-Id: Idc928d7a044886c6474716576adb7cc1993c41f9