Ignore TObjectHash.REMOVED in BLeak tests
TObjectHash.REMOVED is a singleton used to mark removed elements in
Trove collections. When an item is added to and removed from such a
collection within a BLeak iteration, an additional reference to
TObjectHash.REMOVED is added to the collection, which registers as a
leak.

This change prevents the ArrayObjectIdentityExpander from creating edges
to TObjectHash.REMOVED.

Change-Id: I7c5bcf532737ab03a66e567949ddd958037e32e2