Introduce ThrottlingAsyncSupplier
When creating an AsyncSupplier for a value which is frequently
invalidated and also frequently requested (e.g. merged manifests), we'd
like to throttle the recomputation rate so that we recompute the value
no more than once in an N millisecond interval. Otherwise, we run the
risk of wasting resources and holding on to the global read/write lock
just to compute and cache values that are immediately invalidated.

This CL resolves the issue by providing ThrottlingAsyncSupplier, an
implementation of AsyncSupplier which ensures a minimum interval between
computations by delaying the computation and merging any incoming
requests made during that time. The waiting mechanism is powered by
IntelliJ's Alarm, similar to MergingUpdateQueue.

Bug: 128610392
Test: ThrottlingAsyncSupplierTest
Change-Id: I2d89becca7b3036f420f78d3a44c35f134597e63