Remove unnecessary synchronization code in TransportServiceProxy
... which was causing flakiness. The problem was that when the thread
that streams events from daemon was terminating, it needs to wait for
the other thread that streams events to the datastore to finish before
creating all the end events. Instead of adding more synchronization to
make sure that happens, this change mvoes the group ending logic onto
the same thread.

Bug: 126763044
Test: Existing
Change-Id: Id9c248bbc58442458b07ccb6ea0c7b4e418dc100