Enable instant app deployment for CreateAndRunInstantAppTest
The test was not deploying the app as an instant app, so the test was
generating false positives. After this change, the test will generate
real negatives, but the underlying bug can not be fixed. The instant app
library uses its own logic for searching and using ADB rather than using
ddmlib.

Bug: 79937083
Test: N/A
Change-Id: Ib98ea308a49da3f38f373de2d9ead05490ad80c4