Set up LLDB in location the IDE expects during test
When UI tests run within Bazel, the IDE does not consider itself to be
running from sources, at least according to
PluginManagerCore.isRunningFromSources. The Android NDK plugin calls
PluginManagerCore.isRunningFromSources to check what configuration the
IDE is running in. The plugin chooses different filepaths based on the
value of PluginManagerCore.isRunningFromSources.

When running within Bazel, PluginManagerCore.isRunningFromSources
returns false. The NDK plugin then treats the IDE as if it was running
from a normal, user installation. The NDK checks if LLDB is available
under the bin/ directory. In Bazel, this directory is the tools/idea/bin
directory.

Since the test expects the LLDB binaries to be available under
tools/idea/bin. This change sets up tools/idea/bin to contain the LLDB
binaries as the Android NDK plugin expects.

Bug: 129001744
Test: DualDebuggerBreakpointsTest
Change-Id: I872d55359251289c267676ecb74c0db2b9c478b0