Make handling of disposed modules more robust
In the context of RenderTask disposal ModuleBasedClassFileFinder may be called on a pooled thread running concurrently with disposal of projects and its contents. To prevent propagation of exceptions caused by disposed modules, the isDisposed checks are moved to the catch clauses.

This is an attempt to fix the memory leak reported in b/126125131, but it is quite likely that other problems caused by asynchronous disposal of RenderTask might surface.

Test: NavigationViewTest hopefully should not leak memory
Bug: 126125131
Change-Id: I0e51926fda49b3429d13b56b3d06a45cb0098210