Report objects retained by all leak root children
Currently BLeak reports the objects retained by the children added
to a suspected leak root during the final iteration. This can
underestimate the severtity of leaks in which the same heavy objects are
referenced by the leaked objects. This change adds reporting for the
objects retained by all of the leak root's children.

Change-Id: Icb526868030049f4e1f7255cd8d5f7abc7e4f33f