Proper anchor draw modes when making connections
Some modes are not being used, and isTargeted is redundant since
it checks for selected and mouse is over.

So changing to check for doing a connection, is a bit redundant
to can_connect but it's important to make the distinction of
unconnectable anchors WHILE creating constraints.

Anchors should now show a hover ring when creating a constraint to them,
and draw as delete when appropiate.

Fixes: 129712491
Test: n/a
Change-Id: I676cf0bd47144700c1607a5ffea21a908ebe2396