Add keyboard access for opening titles in properties panel.
Before this change a CollapsibleLabel was simple a JBLabel where
the icon in the label was used for expand/collapse icons. That
icon is not focusable. In addition the action buttons were
"slabbed" on externally.

Fix this by making the CollapsibleLabel a JPanel with a focusable
icon, a label, and some optional action buttons.

After this fix the user will be able to <TAB> to a section title
and expand the section with <space> and <enter>.
The action buttons on "Declared Attributes" is accessible though
the keyboard as well. Allthough - is not that usefull from the
keyboard.

Fixes: 129868837
Test: TODO:129977625
Change-Id: Ie06dd327b0faccf23cf75c9e494d7d61f1c76e43