Push the checker validate calls off the EDT
This used to be the LaunchCompatibilityChecker counterpart to 5fa841b2d6f. A
pre submit check failed because LaunchCompatibilityChecker::validate calls
were being made on the EDT.

I extended the original change to push the calls off the EDT. I moved them
from the Device constructor to the virtual and connected device workers. The
connected device worker needed to return both the ddmlib device and the
valid bit from LaunchCompatibilityChecker so I added ConnectedDevice.

Bug: Not applicable
Test: Existing tests
Change-Id: I85425236afef9ac5ac91537ecbbfdfcbeaa3c1fb