Do not scroll the container when attempting to find focus.
When searching for focus the table in the properties panel will
attempt to create an editor in the panel and scroll the cell into
view.

When a table is hidden we should fail early because we know we are
not going to get the focus and we do not want to scroll in that
case.

Fixes: 129978958
Test: TODO:129980950
Change-Id: I1c444261657195001d20caccfe195e90c4595066