Only show apply changes UI for android run configs
Currently the apply changes UI is shown for all types of run
configurations including those that apply changes doesn't apply to.

This CL adds a check in BaseAction to only show apply changes UI if a
relevant android run configuration is selected. E.g. If the selected run
configuration is a java test run configuration, then apply changes UI
won't be shown because it's irrelevant to java tests.

This CL also does a minor clean up of BaseAction.java by extracting out
logic fron update() to isApplyChangesRelevant and
isApplyChangesApplicable.

Fixes: 128685291
Test: Manually verified that apply changes UI only shows for android run
configurations in the IDE.

Change-Id: I5653956c81b1915699aba23dbae329740c6080a1