Make newAttributeRule only match attributes
I assume that the intent of XmlRearranger::attrArrangementRule is to only
match attributes and not elements. Something changed in 3.5 that make those
rules match elements as well. That in turn makes the arrangement engine
rearrange elements when developers reformat their code which changes XML
semantics in bad ways.

Add an AndroidXmlRearranger::newAttributeRule method, add an XML attribute
atom match condition to the rule it initializes, and replace the calls to
the old method with the new one

Bug: 129323345
Test: AndroidXmlArrangementTest::testArrangeDoesntRearrangeElements
Change-Id: I6ecf30a08944ea4bb59ab4491c8077738930d2d9