Fix AndroidExtractionTest for Kotlin 1.3.30
AndroidExtractionTestGenerated still has some dependencies on Kotlin plugin
test code, and it starts to fail when upgrading to Kotlin 1.3.30 due to binary
incompatibilities between 1.3.30 and the current version of the Kotlin test jars
in //prebuilts/tools/common/android-kotlin.

There are generally two options when this happens: (1) upgrade the
Kotlin test jars in tandem with the Kotlin plugin (by doing a custom
build and copying the test jars into prebuilts), or (2) try to reduce
the dependency on the Kotlin test code so that the test passes. This
CL chooses option #2, because that route also helps pay down the
techical debt of depending on Kotlin test code. In this case, option #2
meant copying over some test code and modifying it to be compatible
with 1.3.30.

Test: existing
Bug: n/a
Change-Id: I422323107ca9eb9e3ba2b5c1bc60bd4749f6dd30
(cherry picked from commit 65784f73c9bc8fae0061f413aebc8e80c601c438)