Move namespaces to root tag when adding an attribute.
When the layout editor is adding a non android attribute it will
ensure the app namespace is declared on the root tag in the layout
file.

With the addition of the <layout> tag for data binding people often
have added an empty <layout> tag around an existing layout file
such that the namespace declarations appear on a different tag than
the root tag (the layout).

This CL will move the namespaces out of the tag that is being
changed to the root tag.

Also: fix a bug in the code for moving the namespace. If the root
is updated with an attribute value the existing code would remove
all but the newly added namespace.

Fixes: 120507391
Test: Unit tests added.
Change-Id: I06aee5d210a76466a16e334b92dc5ec98b58b989