Log a warning when the selection time is null
When a developer selects a device with the drop down the time that happens
is supposed to be stored. In some cases (which I haven't been able to
reproduce) that time is later retrieved as null.

Log a warning and return a reasonable default when that happens instead of
crashing

Bug: 129772542
Test: DeviceAndSnapshotComboBoxActionTest
  ::getSelectedDeviceSelectionTimeIsNull
Change-Id: Id15dccaa6ec4accd0d930a76bb2d2e84cab5af7f