Set lastComputation before broadcasting result
Currently, when ThrottlingAsyncSupplier finishes computing a fresh
value, it first sets the scheduled computation's future and then
updates the lastComputation field. Instead, we should set
lastComputation before broadcasting the result so that getNow()
callers get as fresh a value as possible. Doing so also makes the
class easier to test because lastComputation is guaranteed to have
been updated once the future is complete (there's technically a race
condition in ThrottlingAsyncSupplierTest#now_matchesLastComputation).

This CL resolves the issue by splitting up Computation#complete so that
the complete() function simply sets the result and the broadcastResult()
function sets the future. This allows ThrottlingAsyncSupplier to update
lastComputation with a valid Computation before broadcasting the result.

Test: ThrottlingAsyncSupplierTest
Bug: 128610392

Change-Id: Ie1df1cd526843a64559f9f0c40148cca96bc69a0