Don't drop tasks from included build modules.
The tasks from included build modules can be run with the
correct root project directory.
The root directory is in TaskData::myLinkedExternalProjectPath,
which is used as project directory when running the task from
Gradle Task window.
This feature is enabled in Intellij, and this change enables
it for Studio as well.

Bug: NA
Test: AndroidGradleProjectResolverIdeaTest
Change-Id: I0f8ee18e153e34ffa6ec609acabdf07ddc86059f