Ensure the android namespace is present before set attr.
When dragging a component from the palette the layout editor is
assigning an id and setting the attribute with namespace.
However the new component is missing the android namespace
declaration and the added id property will not have a prefix.

Fix this by ensuring the namespace declaration is present before
setting the attribute.

Fixes: 130370114
Test: This fixes a broken ui test:
      ComponentTreeTest.testDropThatOpensDialog

Change-Id: I0d32cf5dcd42541522211a1cf068202f09b70690