Fix bug where dead session's agent is UNATTACHABLE
During agent status refactoring we always set agent status to
UNATTACHABLE when a process dies. This caused dead sessions to always
show "Advanced profile unavailable" banners and disable L3 views for
network and energy profilers.
To fix it, we only set the status to UNATTACHABLE when a process dies
if the agent wasn't previously ATTACHED.

Fixes: 130359658
Test: DeviceProcessPollerTest#testStopUpdateDeadProcesses
Change-Id: I394b115e4b045cc330c5161e4fc5b23b828c1c9c