Write aspects agent violation to log instead of failing
Instead of throwing exception when finding a violation, the agent now
logs it to a file, which will later be parsed so we can know what's
broken.

Throwing exceptions make the failures of the agent too catastrophic. As
we're whitelisting quite a few stack traces, a simple refactoring might
break dozens of tests at once, making other failures in UI tests hard to
be spotted.

Instead, we're gonna simply write this violations to a log and parse it
afterwards in a single dedicated tests, which should fail if there are
any violations. This test should run after all the reliable ones.

Bug: 130264928
Test: UI tests
Change-Id: I12bba8d3d21e4975e4fa7c65470556f60f2ba171