Look for aspects agent failures after running UI tests
Create a unit test to be run after all UI tests of the DEFAULT TestGroup.
This test (AspectsAgentLogTest) parses the aspects agent log and fail in
case there are any violations. The violations are then printed to the output
specifying which tests have failed. Developers can then run these specific
tests locally when fixing the issues.

This approach allows us to fail the reliable UI tests target in case
there are violations to the aspects agent rules without interfering in
the result of individual tests, not causing massive slowdowns and
multiple failures when the rules are violated.

Bug: 130264928
Test: UI tests and AspectsAgentLogTest

Change-Id: Ica459378569d044f373073765cd5d378478d80a4