Display non-excluded paths in anti-virus notification
Add the paths that were not excluded from anti-virus scanning to the
notification, so that users know what remains to be excluded.

Add a parameter to AndroidStudioSystemHealthMonitor.showNotification
controlling whether the balloon has collapsible content. Previously,
MyNotification implemented NotificationFullContent, which disables
collapsing long message contents. Now MyNotification does not implement
this, but has a subclass that does. showNotification selects between the
two. Other callers of showNotification retain their original behavior.

Update the URL for the "Details" action to point to the real page.

Bug: 127513071
Bug: 130229195
Change-Id: I2465ac78bd0248d7023ca932f62b7bcde6dc9c14