Catch NullPointerException generated due to UI changing unexpectedly
The build variants tool window now clears the build variants table and
rebuilds the table when a new build variant is selected. This causes an
error in the BuildVariantToolWindowFixture, which was built under the
assumption that the table would not be emptied and rebuilt. Some change
caused the table to adopt the new behavior, which can cause the UI test
fixtures to end up querying the table while the table is still being
rebuilt. If the underlying fixtures query the table during this period,
the fixtures will throw a NullPointerException (NPE), which causes the
FlavorsExecutionTest to fail.

Bug: N/A
Test: bazel test //tools/adt/idea/android-uitests:FlavorsExecutionTest
Change-Id: I39b38b66a20444407ea00291a8ad1353897ced28