Show JDK location step in custom install only
We are currently showing a first run wizard step to setup JDK location,
this change moves this step to the custom install and adds a message in
the install summary when it detects that JAVA_HOME is not used as JDK
location.

Also, moved the JdkSetupStep behind StudioFlags.NPW_SHOW_JDK_STEP and
added StudiFlags.NPW_FIRST_RUN_SHOW to enable first run wizard to be run
always (to help on developping features related to it).

Bug: 115925397
Test: Existing tests and manualy checked.
Test-Strategy: ui.test.groups=PROJECT_SUPPORT,PROJECT_WIZARD,EDITING,TEST_FRAMEWORK,DEFAULT
Change-Id: I1fe02b34e6a0360b4cf8b60e98c7b3588ca11d7a