Make LocalApkProjTest a bit more stable
The test can fail while waiting for the APK file picker dialog to show
the file tree. For some reason, showing the file tree takes a long time,
so this change updates the test to wait much longer for the file tree to
be ready. This is acceptable because this test only checks if the
critical user journey is broken. If the flow is not broken, then the
test should not be failing.

Bug: N/A
Test: bazel test //tools/adt/idea/android-uitests:LocalApkProjTest
Change-Id: Iad347d6b7644a5138a7bba66f52e55ea299f3014