Assert new file is valid
Based on the other logging it seems that attrs2.xml is
not getting included in the resource repository manager.
Assert that the file is added correctly. Also remove the
write command invocation around the file/class adds. It
is possible that the sync() call was occurring before
the write command completed.

Bug: 123959170

Test: this is it
Change-Id: Ib922bd041b472a7288d45c352e8660d01a946d7f