Stop populating Support Repository node in SDK manager.
Support Repository in SDK manager contains outdated stuff,
and we should not be offering it there.

The relevant quickfixes redirect to gMaven instead, which
is currently the source of truth for support components.

This CL removes the code which takes special care of the
maven packages when populating SDK manager nodes, thus
eliminating them from the UI and reducing their exposure
to the user.

Test: This is a UI cleanup; verified visually.
Bug: Continuation of the cleanup started in b/128825869
Change-Id: I8856bcec23ad1e93c5714c9517c587d31bdb97ae