Clear parameters of statements after they are executed
This allows us to free up any resources used by limited use
queries such as insert payload. If insert payload gets called
when we are done with recording and we do not clear the
resources we leave the size of the payload allocated in memory.
For a 60 second perfetto catpure that is roughly 200mb.

Bug: 124038138
Test: None, due to not having an API to see if parameters are set

Change-Id: Ieefa9bbea14d8036c119673ba846d584c8fa93e4