Prevent infinite restart loops in UI tests
At the beginning of a UI test, there's a check to ensure that the
welcome frame, and nothing else, is showing. If not, the IDE is
restarted and the test begins again. This ensures that previous
tests don't cause the current one to fail because they left windows
up, etc. However, if there's a dialog that shows up at startup for
some reason, this will cause an infinite restart loop.

This change limits each test to a single restart, since this is all that
should ever be required to eliminate detritus from previous tests.

Change-Id: Ifdfbbf26dbab002935f797ae01fdf96546bf6a64