Fix buffer overflow error from IDE trying to display a notification
When the IDE displays a notification, the process executes code in the
libnotify library. While executing through the native libraries, the
process will request opening a pipe. The kernel returns to the process
two file descriptors for both ends of the pipe. libglib will attempt to
add one of the file descriptors to a fd_set to wait for data to be read.
However, fd_set has a size limitation of 1024; file descriptors larger
than 1023 will not be able to be used with fd_set. The file descriptor
returned from the reading end of the pipe can be larger than 1024;
during my debug sessions, the file descriptor was 1309. When the file
descriptor is larger than the fd_set size, the process terminates,
saying that there is a buffer overflow.

Normally, the IDE process does not even call into libnotify, but it was
calling into libnotify during a Bazel test. The underlying reason found
was that the IDE would try to show a system notification if a Gradle
sync failed. Inside Bazel, the file paths are long enough that a CMake
invocation fails, since CMake 3.6 can only read up to 31 file path
components. When the CMake invocation fails, the Gradle sync fails, and
the IDE attempts to show a system notification for the failure. This
causes the process to enter libnotify and eventually encounter the
buffer overflow error.

The solution here is to avoid having the Gradle sync fail in the first
place. To do this, the test project is updated to use CMake 3.10
immediately upon import rather than waiting for the test to
programmatically update the project to use CMake 3.10.

Bug: N/A
Test: bazel test //tools/adt/idea/android-uitests:LocalApkProjTest
Change-Id: I93bc1dfd63e912d50b1a90be8ffef7fe1595924c